refactor: extract video actions from AgentWorkspace
