+++ src/components/agent/AgentWorkspace.tsx
@@ -45,7 +45,6 @@ import {
   assignSessionToProfile,
   listSessionProfiles,
   cancelAgentTask,
-  computerUseBeginRun,
   computerUseEndRun,
   computerUseStop,
   dictationHelperCommand,
@@ -59,7 +58,6 @@ import {
   hermesAgentCliAccess,
   hermesBrowserAccess,
   hermesBridgeSkills,
-  generateImage,
   primeGeneratedVideoDir,
   hermesBridgeStatus,
   hermesBridgeToolsets,
@@ -73,12 +71,9 @@ import {
   providerModelSettings,
   registerBrowserExtensionHost,
   retryAgentTask,
-  imagePromptMayBeExplicit,
   revealPath,
   setHermesAgentCliAccess,
   setHermesBrowserAccess,
-  setImageSafeMode,
-  setImageSafeModePromptDismissed,
   setLocalGenerationEnabled,
   setCostQuality,
   setVeniceModel,
@@ -87,8 +82,6 @@ import {
   toggleHermesBridgeSkill,
   toggleHermesBridgeToolset,
   updateHermesBridgeMessagingPlatform,
-  videoGenerate,
-  videoStatus,
   type AgentTaskDto,
   type AgentTaskStatus,
   type HermesBridgeStatus,
@@ -135,13 +128,8 @@ import {
   cancelAgentRunMonitoring,
   markAgentRunSucceeded,
   releaseAgentRunSettlement,
-  startAgentRunMonitoring,
 } from "../../lib/agent-run-monitor";
-import {
-  HermesGatewayClient,
-  isSessionBusyError,
-  type HermesGatewayEvent,
-} from "../../lib/hermes-gateway";
+import { HermesGatewayClient, type HermesGatewayEvent } from "../../lib/hermes-gateway";
 import {
   classifyHermesEvent,
   createHermesMethods,
@@ -208,7 +196,6 @@ import {
   MODEL_SWITCH_NEXT_MESSAGE_NOTICE,
   MODEL_SWITCH_DEFAULT_ONLY_NOTICE,
 } from "../../lib/hermes-model-switch";
-import { applySessionModelWhenIdle } from "../../lib/hermes-next-prompt-model";
 import {
   reserveHermesSessionDispatch,
   type HermesSessionDispatchReservation,
@@ -218,10 +205,8 @@ import {
   forgetSessionModelSelection,
   hasPendingSessionModelSelection,
   hermesModelIdForSelection,
-  markSessionModelSelectionApplied,
   migrateSessionModelSelection,
   readSessionModelSelections,
-  rememberAppliedSessionModelSelection,
   stageSessionModelSelection,
   subscribeSessionModelSelections,
   type SessionModelSelection,
@@ -264,7 +249,6 @@ import {
   MESSAGING_PLATFORMS_LOAD_TIMEOUT_MESSAGE,
   MESSAGING_PLATFORMS_LOAD_TIMEOUT_MS,
 } from "../../lib/hermes-messaging";
-import { categoryPrompt } from "../../lib/issue-report-prompt";
 import {
   explicitSkillInvocationPrompt,
   isPathLikeSlashToken,
@@ -281,12 +265,6 @@ import {
   resolveSlashModel,
   slashModelResolutionError,
 } from "../../lib/agent-composer-slash-commands";
-import { generateChatImage, newImageRequestId } from "../../lib/chat-image-generation";
-import {
-  generateChatVideo,
-  newVideoRequestId,
-  pollChatVideo,
-} from "../../lib/chat-video-generation";
 import { IMAGE_GENERATION_ENABLED, VIDEO_GENERATION_ENABLED } from "../../lib/feature-flags";
 import { ImageSafeModeConsentDialog } from "./ImageSafeModeConsentDialog";
 import { VideoSafeModeConsentDialog } from "./VideoSafeModeConsentDialog";
@@ -297,11 +275,7 @@ import {
 } from "./composer/ComposerEditor";
 import { noteReferenceToken, type NoteReferenceInput } from "./composer/noteReference";
 import { CategoryIcon } from "./composer/CategoryIcon";
-import {
-  ISSUE_REPORT_ATTACHMENTS_ONLY_DESCRIPTION,
-  REPORT_CATEGORIES,
-  type ReportCategory,
-} from "./composer/reportCategory";
+import { REPORT_CATEGORIES, type ReportCategory } from "./composer/reportCategory";
 import { ReportDialog, type ReportDialogAttachment } from "./ReportDialog";
 import { hermesConnectionForMode } from "../../lib/hermes-connection";
 import {
@@ -323,7 +297,6 @@ import {
 } from "../../lib/agent-chat-runtime";
 import {
   COMPACTED_CONTEXT_SIGNATURE,
-  prepareProjectPrompt,
   ProjectContextSignatureStore,
 } from "../../lib/agent-project-context";
 import {
@@ -333,6 +306,7 @@ import {
 } from "../../lib/agent-chat-gallery";
 import { attachScrollThumbFade } from "../../lib/scroll-thumb-fade";
 import type { AgentWorkspaceProps } from "./agent-workspace-types";
+import { createVideoSlashActions } from "./video-slash-actions";
 import { createImageSlashActions } from "./image-slash-actions";
 import { createSubmitHermesSession } from "./session-submission";
 import type { SubmitHermesSession } from "./session-submission-types";
@@ -435,7 +409,6 @@ export {
 import type { PendingIssueReport } from "./agent-session-continuity";
 
 import type {
-  ImageSafeModeConsentChoice,
   ImageSafeModeConsentEventPayload,
   ImageSafeModeConsentRequest,
 } from "./agent-workspace-models";
@@ -458,29 +431,16 @@ type AgentDeleteSessionDetail = {
 
 import {
   AttachBlockedError,
-  filenameFromWorkspacePath,
   imageSlashTurnsBySessionFromStored,
   markStoredImageSlashTurnsAttached,
-  markStoredVideoSlashContextsSent,
-  promptSubmitContentWithFastPathImageContext,
   removeStoredImageSlashSession,
   removeStoredVideoSlashSession,
-  removeStoredVideoSlashTurn,
   runningImageSlashTurns,
-  runningVideoSlashTurns,
-  storedPendingImageSlashAttachments,
-  storedPendingVideoSlashContexts,
   storedVideoSlashTurns,
-  uniqueAttachmentsByWorkspacePath,
-  upsertStoredImageSlashTurn,
-  upsertStoredVideoSlashTurn,
   videoSlashTurnsBySessionFromStored,
-  withVideoFastPathContext,
-  type PersistedVideoSlashTurn,
 } from "./composer/media-slash-persistence";
 import {
   buildUpNextDemoFollowUps,
-  sameSessionModelSelection,
   type CapturedSessionModelTarget,
   type PendingAttachmentPreparation,
   type PendingSteer,
@@ -489,11 +449,9 @@ import {
 } from "./composer/follow-up-queue";
 
 import {
-  appendIssueReportFollowUp,
   captureSessionContinuity,
   clearAgentSessionContinuity,
   dispatchIssueReportDeliverySettled,
-  dispatchIssueReportFollowUpSubmitFailed,
   issueReportDescription,
   issueReportSentMessage,
   messageAfterIssueReportDiagnosisBoundary,
@@ -3416,426 +3374,29 @@ export function AgentWorkspace({
       : assistantTurnId;
   }
 
-  async function finishVideoSlashGeneration(input: {
-    sessionId: string;
-    turnId: string;
-    prompt: string;
-    requestId: string;
-    createdAt: string;
-    videoCreatedAt: string;
-    model?: string;
-    jobId?: string;
-  }) {
-    const { sessionId, turnId, prompt, requestId, createdAt, videoCreatedAt } = input;
-    const assistantTurnId = `${turnId}:assistant`;
-    try {
-      const result = input.jobId
-        ? await pollExistingVideoSlashJob(input)
-        : await generateChatVideo(
-            prompt,
-            {
-              startGenerate: async (text, model, nextRequestId, options) => {
-                const job = await videoGenerate({
-                  prompt: text,
-                  model,
-                  requestId: nextRequestId,
-                  ...options,
-                });
-                updateVideoSlashPart(sessionId, assistantTurnId, { jobId: job.jobId });
-                upsertStoredVideoSlashTurn({
-                  id: turnId,
-                  sessionId,
-                  prompt,
-                  path: "",
-                  name: "",
-                  createdAt,
-                  videoCreatedAt,
-                  pending: true,
-                  requestId,
-                  model: input.model,
-                  jobId: job.jobId,
-                });
-                return job;
-              },
-              pollStatus: videoStatus,
-              onProgress: (progress) => {
-                updateVideoSlashPart(sessionId, assistantTurnId, {
-                  jobId: progress.jobId,
-                });
-                upsertStoredVideoSlashTurn({
-                  id: turnId,
-                  sessionId,
-                  prompt,
-                  path: "",
-                  name: "",
-                  createdAt,
-                  videoCreatedAt,
-                  pending: true,
-                  requestId,
-                  model: input.model,
-                  jobId: progress.jobId,
-                });
-              },
-            },
-            input.model,
-            requestId,
-            {},
-          );
-      if (result.status !== "ok") {
-        updateVideoSlashPart(sessionId, assistantTurnId, {
-          status: "error",
-          error: result.message,
-          jobId: result.jobId,
-        });
-        if (!result.stillRunning) {
-          removeStoredVideoSlashTurn(turnId);
-        }
-        return;
-      }
-      const name = filenameFromWorkspacePath(result.path, "generated-video.mp4");
-      updateVideoSlashPart(sessionId, assistantTurnId, {
-        status: "complete",
-        path: result.path,
-        name,
-        model: result.model ?? input.model,
-      });
-      upsertStoredVideoSlashTurn({
-        id: turnId,
-        sessionId,
-        prompt,
-        path: result.path,
-        name,
-        createdAt,
-        videoCreatedAt,
-        requestId,
-        model: result.model ?? input.model,
-        jobId: result.jobId,
-        // Hold this turn's context for the video fold: the next real prompt in
-        // this session carries it to the model (storedPendingVideoSlashContexts).
-        contextPending: true,
-      });
-      hermesArtifactStore.recordArtifact(
-        {
-          sessionId,
-          kind: "file",
-          action: "created",
-          path: result.path,
-          displayName: name,
-          previewAvailable: false,
-        },
-        hermesModeFor(sessionId),
-      );
-      void loadFilesystemSnapshot();
-    } catch (err) {
-      updateVideoSlashPart(sessionId, assistantTurnId, {
-        status: "error",
-        error: messageFromError(err),
-      });
-    } finally {
-      setGeneratingVideo(false);
-      setImportingFiles(false);
-    }
-  }
-
-  async function pollExistingVideoSlashJob(input: {
-    sessionId: string;
-    turnId: string;
-    prompt: string;
-    requestId: string;
-    createdAt: string;
-    videoCreatedAt: string;
-    model?: string;
-    jobId?: string;
-  }) {
-    if (!input.jobId) {
-      return { status: "error" as const, message: "Generation was interrupted." };
-    }
-    // Poll the existing job with the full loop (not a single shot) so a retry
-    // follows it to completion, re-attaching to the same server-side job.
-    return pollChatVideo(input.jobId, {
-      pollStatus: videoStatus,
-      onProgress: (progress) => {
-        updateVideoSlashPart(input.sessionId, `${input.turnId}:assistant`, {
-          jobId: progress.jobId,
-        });
-        upsertStoredVideoSlashTurn({
-          id: input.turnId,
-          sessionId: input.sessionId,
-          prompt: input.prompt,
-          path: "",
-          name: "",
-          createdAt: input.createdAt,
-          videoCreatedAt: input.videoCreatedAt,
-          pending: true,
-          requestId: input.requestId,
-          model: input.model,
-          jobId: input.jobId,
-        });
-      },
-    });
-  }
-
-  // Resume a `/video` turn whose poll loop was lost (app crash, restart, or dev
-  // hot-reload). The server job keeps running, so re-attach with the SAME poll
-  // loop and follow it to completion instead of a single shot — the user gets
-  // the video without a new billable generation, and never has to hit "Try
-  // again" just because the app closed mid-render.
-  async function resumePendingVideoSlashTurn(turn: PersistedVideoSlashTurn) {
-    if (!turn.jobId) return;
-    const jobId = turn.jobId;
-    const assistantTurnId = `${turn.id}:assistant`;
-    const result = await pollChatVideo(jobId, {
-      pollStatus: videoStatus,
-      onProgress: (progress) => {
-        updateVideoSlashPart(turn.sessionId, assistantTurnId, {
-          status: "running",
-          jobId: progress.jobId,
-        });
-        upsertStoredVideoSlashTurn({
-          ...turn,
-          pending: true,
-        });
-      },
-    });
-    if (result.status === "ok") {
-      const name = filenameFromWorkspacePath(result.path, "generated-video.mp4");
-      updateVideoSlashPart(turn.sessionId, assistantTurnId, {
-        status: "complete",
-        path: result.path,
-        name,
-        model: result.model ?? turn.model,
-      });
-      upsertStoredVideoSlashTurn({
-        ...turn,
-        pending: false,
-        path: result.path,
-        name,
-        model: result.model ?? turn.model,
-        // Fold this turn's context into the next prompt, same as a live finish.
-        contextPending: true,
-      });
-      hermesArtifactStore.recordArtifact(
-        {
-          sessionId: turn.sessionId,
-          kind: "file",
-          action: "created",
-          path: result.path,
-          displayName: name,
-          previewAvailable: false,
-        },
-        hermesModeFor(turn.sessionId),
-      );
-      void loadFilesystemSnapshot();
-      return;
-    }
-    // Budget exhausted while the job was still processing: it lives on the
-    // server, so keep the turn pending (its stored jobId) and leave the loader
-    // up — the next app launch resumes this exact loop. Only a real Venice
-    // failure or a poll error is terminal and surfaces as retryable.
-    if (result.stillRunning) {
-      updateVideoSlashPart(turn.sessionId, assistantTurnId, {
-        status: "running",
-        jobId,
-      });
-      return;
-    }
-    updateVideoSlashPart(turn.sessionId, assistantTurnId, {
-      status: "error",
-      error: result.message,
-      jobId,
-    });
-    removeStoredVideoSlashTurn(turn.id);
-  }
-
-  async function retryVideoSlashTurn(
-    sessionId: string,
-    assistantTurnId: string,
-    part: Extract<AgentChatPart, { type: "video" }>,
-  ) {
-    if (creditActionsDisabledReason && !part.jobId) {
-      setError(creditActionsDisabledReason);
-      return;
-    }
-    if (part.status !== "error" || !part.requestId) return;
-    const now = new Date().toISOString();
-    setError(null);
-    setImportingFiles(true);
-    setGeneratingVideo(true);
-    updateVideoSlashPart(sessionId, assistantTurnId, {
-      status: "running",
-      error: undefined,
-    });
-    await finishVideoSlashGeneration({
-      sessionId,
-      turnId: videoSlashBaseTurnId(assistantTurnId),
-      prompt: part.prompt,
-      requestId: part.requestId,
-      createdAt: part.userCreatedAt ?? now,
-      videoCreatedAt: part.videoCreatedAt ?? now,
-      model: part.model,
-      jobId: part.jobId,
-    });
-  }
-
-  async function runVideoSlashCommand(
-    argument: string,
-    commandText: string,
-    modelTarget = captureSessionModelTarget(),
-    dispatchReservation?: HermesSessionDispatchReservation,
-  ) {
-    if (creditActionsDisabledReason) {
-      setError(creditActionsDisabledReason);
-      return;
-    }
-    const prompt = argument.trim();
-    if (!prompt) {
-      setError("Type a description after /video to generate a video.");
-      return;
-    }
-
-    // Busy-gate the consent + generation flow before any async IPC, mirroring
-    // /image: a second submission can't start while the prompt screen or
-    // consent dialog is pending, and dismiss leaves the draft untouched.
-    setImportingFiles(true);
-
-    // Pin the video model before the paid turn starts (same replay-ledger
-    // rationale as /image). Safe mode is read alongside but never pinned into
-    // the request: video requests carry no safeMode field (Venice cannot blur
-    // video), so the value only gates the consent dialog below.
-    let settings: ProviderModelSettingsDto | undefined;
-    let pinnedModel: string | undefined;
-    try {
-      const settingsResponse = await providerModelSettings();
-      settings = settingsResponse.settings;
-      pinnedModel =
-        settingsResponse.effectiveSettings?.videoModel || settings.videoModel || undefined;
-    } catch {
-      // Non-fatal: generation proceeds with server-resolved settings.
-    }
-
-    // Unlike /image, the screen runs even after "don't ask again": for video
-    // the dialog is the enforcement point (there is no blur to fall back to),
-    // so an explicit prompt with safe mode on must never generate silently.
-    if (settings?.imageSafeMode) {
-      let mayBeExplicit = false;
-      try {
-        mayBeExplicit = await imagePromptMayBeExplicit(prompt);
-      } catch {
-        mayBeExplicit = false;
-      }
-      if (mayBeExplicit) {
-        if (settings.imageSafeModePromptDismissed) {
-          // The user opted out of the dialog, not out of safe mode: skip the
-          // generation with a notice instead of asking again.
-          setImportingFiles(false);
-          setError(
-            "Safe mode is on, so this video was skipped. Turn safe mode off in Settings to generate it.",
-          );
-          return;
-        }
-        const choice = await requestImageSafeModeConsent("video-slash", dispatchReservation);
-        if (choice.action === "dismiss") {
-          setImportingFiles(false);
-          return;
-        }
-        if (choice.action === "keep") {
-          // "Skip this video": no blurred fallback exists for video, so safe
-          // mode on means the generation is skipped (the dialog says so).
-          if (choice.dontAskAgain) void setImageSafeModePromptDismissed(true);
-          setImportingFiles(false);
-          return;
-        }
-        try {
-          await setImageSafeMode(false);
-        } catch (err) {
-          setImportingFiles(false);
-          setError(messageFromError(err));
-          return;
-        }
-        if (choice.dontAskAgain) void setImageSafeModePromptDismissed(true);
-      }
-    }
-
-    if (composerDispatchWasInvalidated(dispatchReservation)) {
-      setImportingFiles(false);
-      return;
-    }
-
-    const heroMode = newSessionModeRef.current;
-    if (heroMode) setHeroLeaving(true);
-    clearComposerCommandDraft(commandText);
-    setError(null);
-    setGeneratingVideo(true);
-
-    let targetSessionId: string | undefined;
-    try {
-      targetSessionId = await submitHermesSession(prompt, undefined, {
-        skipPrompt: true,
-        displayContent: prompt,
-        titleContent: prompt,
-        modelTarget,
-        dispatchReservation,
-      });
-    } catch (err) {
-      if (heroMode) setHeroLeaving(false);
-      setGeneratingVideo(false);
-      setImportingFiles(false);
-      setError(messageFromError(err));
-      return;
-    }
-    if (!targetSessionId) {
-      if (heroMode) setHeroLeaving(false);
-      setGeneratingVideo(false);
-      setImportingFiles(false);
-      setError("Could not start a video session. Try again.");
-      return;
-    }
-    const sessionId = targetSessionId;
-
-    const turnStartedAt = Date.now();
-    const turnId = `video:${sessionId}:${turnStartedAt}`;
-    const createdAt = new Date(turnStartedAt).toISOString();
-    const videoCreatedAt = new Date(turnStartedAt + 1).toISOString();
-    const requestId = newVideoRequestId();
-
-    setVideoTurnsBySession((current) => ({
-      ...current,
-      [sessionId]: [
-        ...(current[sessionId] ?? []),
-        ...runningVideoSlashTurns({
-          id: turnId,
-          prompt,
-          requestId,
-          createdAt,
-          videoCreatedAt,
-          model: pinnedModel,
-        }),
-      ],
-    }));
-
-    upsertStoredVideoSlashTurn({
-      id: turnId,
-      sessionId,
-      prompt,
-      path: "",
-      name: "",
-      createdAt,
-      videoCreatedAt,
-      pending: true,
-      requestId,
-      model: pinnedModel,
-    });
-
-    await finishVideoSlashGeneration({
-      sessionId,
-      turnId,
-      prompt,
-      requestId,
-      createdAt,
-      videoCreatedAt,
-      model: pinnedModel,
-    });
-  }
+  const {
+    finishVideoSlashGeneration,
+    pollExistingVideoSlashJob,
+    resumePendingVideoSlashTurn,
+    retryVideoSlashTurn,
+    runVideoSlashCommand,
+  } = createVideoSlashActions({
+    captureSessionModelTarget,
+    clearComposerCommandDraft,
+    composerDispatchWasInvalidated,
+    creditActionsDisabledReason,
+    loadFilesystemSnapshot,
+    newSessionModeRef,
+    requestImageSafeModeConsent,
+    setError,
+    setGeneratingVideo,
+    setHeroLeaving,
+    setImportingFiles,
+    setVideoTurnsBySession,
+    submitHermesSession,
+    updateVideoSlashPart,
+    videoSlashBaseTurnId,
+  });
 
   if (testOnlySlashCommandEntriesRef) {
     testOnlySlashCommandEntriesRef.current = {
@@ -9354,10 +8915,8 @@ import {
   assignArtifactsToTurns,
   composerInputSignatureFor,
   formatComposerTokenCount,
-  oversizedComposerInputWarning,
   promptWithAttachments,
   surfacedArtifactsFromTurns,
-  unsupportedImageInputPrompt,
   type ComposerInputSizeWarning,
 } from "./composer/composer-input-helpers";
 export { generatedImagePathAliases } from "./composer/composer-input-helpers";
@@ -9388,7 +8947,6 @@ import {
   moveRecordKey,
   omitRecordKey,
   readFileBytes,
-  sameAgentAttachments,
 } from "./agent-workspace-support";
 import {
   AUTO_SUBMIT_ECHO_WINDOW_MS,
